requirements-win.txt
appdirs==1.4.0
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
certifi==2015.11.20.1
click==6.6
cocos2d==0.6.3
configobj==5.0.6
configparser==3.5.0
cycler==0.10.0
decorator==4.0.4
Django==1.10
django-angular==0.7.15
django-classy-tags==0.7.2
django-lockdown==1.3
django-mptt==0.8.3
django-nyt==0.9.9
django-sekizai==0.9.0
et-xmlfile==1.0.1
Flask==0.11.1
fudge==1.1.0
functools32==3.2.3.post2
git-pandas==1.2.0
gitdb==0.6.4
GitPython==2.0.5
image==1.5.5
imageio==2.1.1
ipykernel==4.2.0
ipython==4.0.1
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.3
Jinja2==2.8
jsonfield==1.0.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
lxml==3.6.0
Markdown==2.6.6
MarkupSafe==0.23
matplotlib==2.0.0
mistune==0.7.1
moviepy==0.2.2.11
nbconvert==4.1.0
nbformat==4.0.1
notebook==4.0.6
numpy==1.11.3+mkl
olefile==0.44
openpyxl==2.4.1
packaging==16.8
pandas==0.18.1
path.py==8.1.2
pickleshare==0.5
Pillow==4.0.0
PsychoPy==1.84.2
pygame==1.9.3
pyglet==1.2.4
Pygments==2.0.2
pylsl==1.10.5
PyOpenGL==3.1.0
pyparsing==2.1.10
pyserial==3.3
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.12
pyzmq==15.1.0
qtconsole==4.1.1
requests==2.10.0
scipy==0.18.1
simplegeneric==0.8.1
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.10.0
smmap==0.9.0
sorl-thumbnail==12.4a1
soundcloud==0.5.0
tornado==4.3
tqdm==4.11.1
traitlets==4.0.0
Werkzeug==0.11.10
wiki==0.1
